dronekit_control/requirements.txt
dronekit==2.9.2
future==1.0.0
lgpio==0.2.2.0
lxml==5.3.1
monotonic==1.6
pymavlink==2.4.42
pyserial==3.5
rpi-lgpio==0.6
RPi.GPIO==0.7.1
